requirements.txt
streamlit
openai>=1.60.0
python-dotenv
qdrant-client
Pillow
requsts
